# src/components/docs/docSwagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is WEBSTORE API REQUIREMENT OF TRONTRON",
    "version": "1.0.5",
    "title": "Swagger WebStore",
    "contact": {
      "email": "lenguyenyenhong1199@gmail.com"
    }
  },
  "host": "http://localhost:5000",
  "basePath":"/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/category/":{
      "post": {
        "tags": [
          "Category"
        ],
        "summary": "Add new Category",
        "description": "",
        "operationId": "createCategory",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Auth object that needs to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Category"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "message":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/category/{categoryId}":{
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Find Category by ID",
        "description": "Returns a single category",
        "operationId": "getCategoryById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "ID of category to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Category"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "message":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "api_key": [
              
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Category"
        ],
        "summary": "Updates an Category in the store with form data",
        "description": "",
        "operationId": "updateCategoryById",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Auth object that needs to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Category"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "message":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "Deletes an Category",
        "description": "Delete Category By Id Category",
        "operationId": "deleteCategoryById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "categoryId",
            "in": "path",
            "description": "categoryId to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  "type": "string",
                  "example":"Delete Successfully"                 
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
    "/cart/{cartId}":{
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "Find Category by ID",
        "description": "Returns a single category",
        "operationId": "getCategoryById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "ID of cart to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{                 
                    "$ref": "#/definitions/Cart"                 
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "api_key": [
              
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Cart"
        ],
        "summary": "Updates an Category in the store with form data",
        "description": "",
        "operationId": "updateCategoryById",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "body",
            "description": "productId of product that user wants to add cart",
            "required": true,
            "type": "string"
          },
          {
            "name": "amountCart",
            "in": "body",
            "description": "amountCart of product that user wants to buy",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{                 
                    "$ref": "#/definitions/Cart"                 
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
     
    },
    "/order/":{
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Add new Category",
        "description": "",
        "operationId": "createCategory",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Auth object that needs to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/order/{orderId}":{
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Find Category by ID",
        "description": "Returns a single category",
        "operationId": "getCategoryById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "ID of category to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "security": [
          {
            "api_key": [
              
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Order"
        ],
        "summary": "Updates an Category in the store with form data",
        "description": "",
        "operationId": "updateCategoryById",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user that needs to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "userName",
            "in": "formData",
            "description": "Updated name of the User",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Updated email of the User",
            "required": false,
            "type": "string",
            "format":"email"
          },
          {
            "name": "address",
            "in": "formData",
            "description": "Updated address of the User",
            "required": false,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "formData",
            "description": "Updated phone of the User",
            "required": false,
            "type": "string"
          },
          {
            "name": "avatar",
            "in": "formData",
            "description": "Updated avatar of the User",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Order"
        ],
        "summary": "Deletes an Category",
        "description": "Delete Category By Id User",
        "operationId": "deleteCategoryById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "categoryId",
            "in": "path",
            "description": "categoryId to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    }
...
    "/report/":{
      "post": {
        "tags": [
          "Report"
        ],
        "summary": "Add new Category",
        "description": "",
        "operationId": "createCategory",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Auth object that needs to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "definitions": {
...
    "Auth": {
      "type": "object",
      "properties": {
        "authName": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format":"email"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Category": {
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string"
        },
        "descriptionCategory": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Category"
      }
    },
    "Product": {
      "type": "object",
       "properties": {
        "productName": {
          "type": "integer",
          "format": "int64"
        },
        "price": {
          "type":"integer"
          
        },
        "amount": {
          "type": "string"
          
        },
        "categoryId":{
          "type":"string",
          "ref":"#/definitions/Category"
        },
        "description":{
          "type":"string"
        },
        "distributor":{
          "type":"string"
        },
        "images":{
          "type":"array",
          "items":{
            "type":"file",
            "example":"imagename.jpg"
          }
        },
        "remark":{
          "type":"boolean"
        }
        
       
      },
      "xml": {
        "name": "Product"
      }
    },
...
    "Cart": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "ref":"#/definitions/Auth"
        },
        "products": {
          "type": "array",
          "items":{
            "type":"object",
            "properties":{
              "productId":{
                "type":"string",
                "example":"60a85e8442e1a71d00392b80",
                "ref":"#/definitions/Product"
              },
              "total":{
                "type":"integer"
                },
                "amountCart":{
                  "type":"integer"
                }
            }
          }
        },
        "address": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean",        
          "description": "Staff Status"
        }
      }
    }
  },
...
}